refactoring + BindChecker for projects+ strict check of path parameters
