+++ src/main/java/com/alkl1m/taskmanager/util/checker/BindingChecker.java
@@ -1,4 +1,4 @@
-package com.alkl1m.taskmanager.aspects;
+package com.alkl1m.taskmanager.util.checker;
 
 import org.springframework.stereotype.Component;
