+++ src/main/java/com/alkl1m/taskmanager/util/checker/BindingCheckerImpl.java
@@ -1,4 +1,4 @@
-package com.alkl1m.taskmanager.aspects;
+package com.alkl1m.taskmanager.util.checker;
 
 
 
@@ -11,7 +11,7 @@ import org.springframework.validation.BindingResult;
 @Aspect
 @Component
 public class BindingCheckerImpl {
-    @Before(value = "@annotation(BindingChecker)" + " && args(bindingResult)", argNames = "bindingResult")
+    @Before(value = "@annotation(com.alkl1m.taskmanager.util.checker.BindingChecker)" + " && args(bindingResult)", argNames = "bindingResult")
     public void checkBindingErrors(BindingResult bindingResult) throws BindException {
         if (bindingResult.hasErrors()) {
             if (bindingResult instanceof BindException exception) {
